==> sub-clustering/Myeloid/01_subset.ipynb <==
suppressMessages(library(ArchR))
suppressMessages(library(Seurat))
suppressMessages(library(Signac))
suppressMessages(library(cowplot))
suppressMessages(library(ggplot2))
suppressMessages(library(dplyr))

# set parameters
set.seed(42)
addArchRThreads(threads = parallel::detectCores() - 2)
addArchRGenome("hg38")

proj <- loadArchRProject("../../../snATAC/DataIntegration/data/VisiumHeart", showLogo = FALSE)

df <- read.csv("../../../snATAC/DataIntegration/data/VisiumHeart/metadata.csv",
              row.names = 1)

head(df)

unique(df$cell_type)

df_sub <- subset(df, cell_type == "Myeloid")

nrow(df_sub)

df_cell_count <- df_sub %>%
    group_by(Sample) %>%
    summarise(count = n()) %>%
    arrange(desc(count)) %>%
    subset(count > 50)

df_cell_count

# we dont want to include the samples with too few cells
df_sub <- subset(df_sub, Sample %in% df_cell_count$Sample)

proj <- subsetArchRProject(proj, 
                           cells = rownames(df_sub),
                           outputDirectory = "../data/snATAC",
                           force = TRUE)
proj <- addTileMatrix(proj,
                      force = TRUE)

proj <- addGeneScoreMatrix(proj,
                           force = TRUE)

## save peak matrix
peakMatrix <- getMatrixFromProject(proj,
                                   useMatrix = "PeakMatrix")

peak_counts <- peakMatrix@assays@data$PeakMatrix
df_rangers <- as.data.frame(peakMatrix@rowRanges@ranges)

rownames(peak_counts) <- paste(peakMatrix@rowRanges@seqnames,
                          df_rangers$start,
                          df_rangers$end,
                          sep = "_") 

saveRDS(peak_counts, file = "../data/snATAC/PeakMatrix.Rds")

## save gene matrix
atac <- getMatrixFromProject(ArchRProj = proj,
                             useMatrix = "GeneScoreMatrix")

gene_counts <- atac@assays@data$GeneScoreMatrix
rownames(gene_counts) <- atac@elementMetadata$name

saveRDS(gene_counts, file = "../data/snATAC/GeneScoreMatrix.Rds")

saveArchRProject(ArchRProj = proj, 
                 load = FALSE)

sessionInfo()
