==> sub-clustering/Myeloid/10_annotation.ipynb <==
suppressMessages(library(ArchR))
suppressMessages(library(Seurat))
suppressMessages(library(Signac))
suppressMessages(library(harmony))
suppressMessages(library(dplyr))
suppressMessages(library(cowplot))
suppressMessages(library(harmony))
suppressMessages(library(Nebulosa))
suppressMessages(library(ggpubr))
suppressMessages(library(Ipaper))
suppressMessages(library(phateR))
suppressMessages(library(ComplexHeatmap))

set.seed(42)
getwd()

coembed <- readRDS("../data/coembed/coembed.cleaning.Rds")
coembed

## visualize known markers
ctrl <- 5

coembed <- AddModuleScore(coembed, features= list(c("LILRA1", "APOBEC3A", "FCN1",
                                            "IFI30", "CDC42EP2", "LST1", "SERPINA1", "CFP", "G0S2", "LILRA5")),
                             name = "CD16_Mo", ctrl = ctrl)

coembed <- AddModuleScore(coembed, features= list(c("PLXDC2",
                                            "SLC35F1", "IGSF21", "PID1", "SLC8A1-AS1", "ADAM28", "MKRN3", "SLC8A1")) ,
                             name = "DOCK4", ctrl = ctrl)

coembed <- AddModuleScore(coembed, features= list(c("F13A1", "NEURL2", "PLTP", "LYVE1", "C1QB",
                                                      "SCN9A", "C1QC", "C1QA", "DAAM2", "FOLR2")),
                             name = "LYVE1_Macro1", ctrl = ctrl)

coembed <- AddModuleScore(coembed, features= list(c("F13A1", "FGF13", "RBPJ", "APOD", "STAB1",
                                                      "CLEC4G", "NEURL2", "SCN9A")),
                             name = "LYVE1_Macro2", ctrl = ctrl)

coembed <- AddModuleScore(coembed, features= list(c("FCER1A", "MAMDC2-AS1", "DPYSL3", "AC092691.1",
                                                   "WDR49", "AC010127.1", "MAMDC2", "LSAMP", "LRRC4C", "LINC01266")),
                             name = "LYVE1_Macro3", ctrl = ctrl)

coembed <- AddModuleScore(coembed, features= list(c("S100A9", "LYZ", "S100A8", "VCAN",
                                                   "S100A12", "FCN1", "CTSS", "COTL1", "TIMP1", "GAPDH")),
                             name = "Mo_classic", ctrl = ctrl)

coembed <- AddModuleScore(coembed, features= list(c("S100A9", "LYZ", "S100A8", "VCAN",
                                                   "S100A12", "FCN1", "CTSS", "COTL1", "TIMP1", "GAPDH")),
                             name = "Mo_classic", ctrl = ctrl)

coembed <- AddModuleScore(coembed, features= list(c("SAT1", "S100A8", "LYZ", "MT-CO2",
                                                   "S100A9", "MT-CO1", "LST1", "FTL", "COTL1", "MT-CO3")),
                             name = "Mo_pi", ctrl = ctrl)

coembed <- AddModuleScore(coembed, features= list(c("SLC16A10", "FSIP1", "AC100849.1", "DOCK4",
                                                   "DNMBP", "PPARD", "MB21D2", "TNFRSF10D", "IL1RAP", "MOB3B")),
                             name = "DOCK4_2", ctrl = ctrl)

coembed <- AddModuleScore(coembed, features= list(c("CPA3", "AC092979.1", "KIT", "CDK15",
                                                   "NTM", "SLC24A3", "KIAA1549", "LINC02147", "MEIS2", "IL18R1")),
                             name = "Mast", ctrl = ctrl)

coembed <- AddModuleScore(coembed, features= list(c("CCL13", "CCL18", "RNASE1", "C1QA",
                                                   "C1QB", "C1QC", "NUPR1", "FXYD6", "SELENOP", "DAB2")),
                             name = "mod", ctrl = ctrl)

coembed <- AddModuleScore(coembed, features= list(c("FABP5", "FTH1", "NPC2", "CTSB",
                                                   "FTL", "PSAP", "CD68", "VIM", "CTSD", "CD63")),
                             name = "AgP", ctrl = ctrl)

coembed <- AddModuleScore(coembed, features= list(c("ALCAM", "LPL", "MITF", "TPRG1",
                                                   "MYO1E", "CCDC88A", "ITGAX", "ARHGAP10", "EPB41L2", "MAP4K3")),
                             name = "IL17RA_Mo", ctrl = ctrl)

options(repr.plot.height = 8, repr.plot.width = 24)

p1 <- plot_density(coembed, feature = "CD16_Mo1", reduction="umap_harmony_v2") + ggtitle("CD16_Mo")
p2 <- plot_density(coembed, feature = "DOCK41", reduction="umap_harmony_v2") + ggtitle("DOCK4")
p3 <- plot_density(coembed, feature = "LYVE1_Macro11", reduction="umap_harmony_v2") + ggtitle("LYVE1_Macro1")
p4 <- plot_density(coembed, feature = "LYVE1_Macro21", reduction="umap_harmony_v2") + ggtitle("LYVE1_Macro2")
p5 <- plot_density(coembed, feature = "LYVE1_Macro31", reduction="umap_harmony_v2") + ggtitle("LYVE1_Macro3")
p6 <- plot_density(coembed, feature = "Mo_classic1", reduction="umap_harmony_v2") + ggtitle("Mo_classic")
p7 <- plot_density(coembed, feature = "Mo_pi1", reduction="umap_harmony_v2") + ggtitle("Mo_pi")
p8 <- plot_density(coembed, feature = "DOCK4_21", reduction="umap_harmony_v2") + ggtitle("DOCK4_2")
p9 <- plot_density(coembed, feature = "Mast1", reduction="umap_harmony_v2") + ggtitle("Mast")
p10 <- plot_density(coembed, feature = "mod1", reduction="umap_harmony_v2") + ggtitle("mod")
p11 <- plot_density(coembed, feature = "AgP1", reduction="umap_harmony_v2") + ggtitle("AgP")
p12 <- plot_density(coembed, feature = "IL17RA_Mo1", reduction="umap_harmony_v2") + ggtitle("IL17RA_Mo")

p <- patchwork::wrap_plots(list(p1, p2, p3, p4, p5, p6, 
                                p7, p8, p9, p10, p11, p12), ncol = 6)

p

options(repr.plot.height = 8, repr.plot.width = 22)
resolutions <- seq(0.1, 1, 0.1)

plotlist <- lapply(resolutions, function(x){
    cols <- ArchR::paletteDiscrete(coembed@meta.data[, glue::glue("RNA_snn_res.{x}")])
    
    p <- DimPlot(coembed, group.by = glue::glue("RNA_snn_res.{x}"), label = TRUE,
             reduction = "umap_harmony_v2", shuffle = TRUE) +
    scale_color_manual(values = cols) +
    xlab("UMAP1") + ylab("UMAP2")
    
    p
})

p <- patchwork::wrap_plots(plotlist, nrow = 2)

p

library(clustree)

options(repr.plot.height = 8, repr.plot.width = 12)

clustree(coembed, prefix = "RNA_snn_res.")

options(repr.plot.height = 5, repr.plot.width = 5)

ht_list <- lapply(resolutions, function(x){
    df <- as.data.frame(coembed@meta.data) %>%
    group_by(across(glue::glue("RNA_snn_res.{x}"))) %>%
    summarise(CD16_Mo = mean(CD16_Mo1),
             DOCK4 = mean(DOCK41),
             LYVE1_Macro1 = mean(LYVE1_Macro11),
             LYVE1_Macro2 = mean(LYVE1_Macro21),
             LYVE1_Macro3 = mean(LYVE1_Macro31),
             Mo_classic = mean(Mo_classic1),
             Mo_pi = mean(Mo_pi1),
             DOCK4_2 = mean(DOCK4_21),
             Mast = mean(Mast1),
             mod = mean(mod1),
             AgP = mean(AgP1),
             IL17RA_Mo = mean(IL17RA_Mo1)) %>%
    textshape::column_to_rownames(glue::glue("RNA_snn_res.{x}"))
    
    df <- scale(df)
    
    ht <- Heatmap(as.matrix(t(df)),
                 name = "z-score",
                 column_title = glue::glue("res.{x}"),
                 clustering_distance_columns = "pearson",
                 clustering_distance_rows = "pearson",
                 clustering_method_columns = "ward.D2",
                 clustering_method_rows = "ward.D2")

    draw(ht)
})

# harmony_embedding <- coembed.sub@reductions$harmony@cell.embeddings

# # run PHATE
# phate_embedding <- phate(harmony_embedding, n.jobs = -1, npca = NULL,
#                         knn.dist.method = "cosine")

# colnames(phate_embedding$embedding) <- c("PHATE_1", "PHATE_2")

# coembed.sub[['phate']] <- CreateDimReducObject(embeddings = phate_embedding$embedding)

# cols <- ArchR::paletteDiscrete(coembed.sub@meta.data[, glue::glue("RNA_snn_res.0.5")])

# options(repr.plot.height = 6, repr.plot.width = 6)

# DimPlot(coembed.sub, group.by = "RNA_snn_res.0.5", label = TRUE,
#              reduction = "phate", shuffle = TRUE) +
#     scale_color_manual(values = cols) +
#     xlab("PHATE1") + ylab("PHATE2")

## add annotation
new.cluster.ids <- c("0" = "LYVE_FOLR_Macrophages",
               "1" = "LYVE_PLTP_Macrophages",
               "2" = "Monocytes",
               "3" = "SPP1_Macrophages",
               "4" = "CCL18_Macrophages")

Idents(coembed) <- "RNA_snn_res.0.4"

coembed <- RenameIdents(coembed, new.cluster.ids)
coembed$annotation <- Idents(coembed)


cols.states <- ArchR::paletteDiscrete(coembed@meta.data[, "annotation"])

p <- DimPlot(coembed, group.by = "annotation", label = FALSE,
             reduction = "umap_harmony_v2", shuffle = TRUE) +
    scale_color_manual(values = cols.states) +
    xlab("UMAP1") + ylab("UMAP2")

options(repr.plot.height = 6, repr.plot.width = 8)

p

options(repr.plot.width = 12, repr.plot.height = 5)

p <- DimPlot(coembed, group.by = "annotation", label = TRUE,
             reduction = "umap_harmony_v2", shuffle = TRUE, split.by = "tech", cols=cols.states) +
    xlab("UMAP1") + ylab("UMAP2")

p

meta.data <- as.data.frame(coembed@meta.data)

df_plot <-  meta.data %>%
    group_by(patient, annotation) %>%
    summarise(counts = n()) %>%
    mutate(cell_proportion = counts / sum(counts))

p1 <- ggplot(df_plot, aes(patient, cell_proportion, fill=annotation)) + 
    geom_bar(stat="identity", position = position_stack(reverse = TRUE)) +
    scale_fill_manual(values = cols.states) +
    theme_cowplot() +
    xlab("") + ylab("") +
    theme(axis.text.x = element_text(angle = 60, hjust = 1))


options(repr.plot.width = 10, repr.plot.height = 6)

p1

df_plot <- meta.data %>%
    group_by(region, annotation) %>%
    summarise(counts = n()) %>%
    mutate(cell_proportion = counts / sum(counts))

p1 <- ggplot(df_plot, aes(region, cell_proportion, fill=annotation)) + 
    geom_bar(stat="identity", position = position_stack(reverse = TRUE)) +
    scale_fill_manual(values = cols.states) +
    theme_cowplot() +
    xlab("") + ylab("") +
    theme(axis.text.x = element_text(angle = 60, hjust = 1))


options(repr.plot.width = 10, repr.plot.height = 6)

p1

df_plot <- meta.data %>%
    group_by(patient_group, annotation) %>%
    summarise(counts = n()) %>%
    mutate(cell_proportion = counts / sum(counts))

p1 <- ggplot(df_plot, aes(patient_group, cell_proportion, fill=annotation)) + 
    geom_bar(stat="identity", position = position_stack(reverse = TRUE)) +
    scale_fill_manual(values = cols.states) +
    theme_cowplot() +
    xlab("") + ylab("") +
    theme(axis.text.x = element_text(angle = 60, hjust = 1))


options(repr.plot.width = 6, repr.plot.height = 6)

p1

df_plot <- meta.data %>% 
            group_by(patient_region_id, annotation) %>%
            summarise(counts = n()) %>%
            mutate(proportion = counts / sum(counts))

df_anno <- coembed@meta.data %>%
    as.data.frame() %>%
    subset(., select = c("patient_region_id", "patient_group")) %>%
    unique()

df_plot <- merge.data.frame(df_plot, df_anno)
    
    p <- ggplot(data = df_plot, aes(x = patient_group, y = proportion)) +
    geom_boxplot2(aes(color = patient_group)) +
    geom_point(aes(color = patient_group), position = position_jitterdodge(), size = 3) +
    facet_wrap(~annotation, nrow = 1) +
    stat_compare_means(comparisons = list(c("group_1", "group_2"),
                                         c("group_2", "group_3"),
                                         c("group_1", "group_3")),
                   method = "wilcox.test") +
    theme_cowplot() +
    xlab("") + ylab("") +
    theme(axis.text.x = element_blank(),
         legend.title = element_blank())

options(repr.plot.width = 15, repr.plot.height = 5)

print(p)

Idents(coembed) <- "annotation"

all.markers <- FindAllMarkers(coembed, only.pos = TRUE, min.pct = 0.25, logfc.threshold = 0.25)
    
df <- all.markers %>%
    group_by(cluster) %>%
    slice_max(n = 10, order_by = avg_log2FC)

p <- DotPlot(coembed, features = unique(df$gene)) + RotatedAxis()
    
options(repr.plot.height = 5, repr.plot.width = 18)

print(p)

if(!dir.exists("../data/coembed/MarkerGenesAfterAnnotation")){
    dir.create("../data/coembed/MarkerGenesAfterAnnotation")
}

markerList <- split(all.markers, all.markers$cluster)
    
    for(i in 1:length(markerList)){
        markerList[[i]] <- markerList[[i]][order(-markerList[[i]]$avg_log2FC), ]
    }
    
    WriteXLS::WriteXLS(markerList,
                   ExcelFileName = glue::glue("../data/coembed/MarkerGenesAfterAnnotation/annotation.xlsx"),
                   SheetNames = names(markerList))

saveRDS(all.markers, glue::glue("../data/coembed/MarkerGenesAfterAnnotation/annotation.rds"))

saveRDS(coembed, file = "../data/coembed/coembed.annotation.Rds")

sessionInfo()
